integrated sidebar with profile components
