+++ client/src/components/profile/Profile.js
@@ -1,13 +1,14 @@
 import React from 'react';
 import Sidebar from './Sidebar'; 
 import RightDashboard from './RightDashboard'; 
+import PredefinedTransactions from './PredefinedTransactions';
 
 const Profile = () => {
     return (
         <div className="profile-container">
             <Sidebar/>
-            <RightDashboard/>
-
+            {/* <RightDashboard/> */}
+            <PredefinedTransactions/>
         </div>
     );
 }
